fix: added todo, moved return element to component
