+++ src/components/frame/v5/pages/BalancePage/partials/BalanceTable/Filters/partials/filters/TokenFilters/TokenFilters.tsx
@@ -1,15 +1,26 @@
 import React, { type FC } from 'react';
 
+import TokenIcon from '~shared/TokenIcon/TokenIcon.tsx';
 import { formatText } from '~utils/intl.ts';
 import Checkbox from '~v5/common/Checkbox/index.ts';
 
 import { useFiltersContext } from '../../../FilterContext/FiltersContext.tsx';
 
-import { useGetTokenTypeFilters } from './consts.tsx';
+import { useGetTokenTypeFilters } from './hooks.ts';
 
 const TokenFilters: FC = () => {
   const { tokenTypes, handleTokenTypesFilterChange } = useFiltersContext();
   const tokenTypesFilters = useGetTokenTypeFilters();
+  const tokenItems = tokenTypesFilters.map(({ token }) => ({
+    symbol: token.symbol,
+    label: (
+      <div className="flex items-center gap-2">
+        <TokenIcon token={token} size="xxxs" />
+        {token.symbol}
+      </div>
+    ),
+    name: token?.tokenAddress || '',
+  }));
 
   return (
     <div>
@@ -17,7 +28,7 @@ const TokenFilters: FC = () => {
         {formatText({ id: 'balancePage.filter.approvedTokenTypes' })}
       </h5>
       <ul>
-        {tokenTypesFilters.map(({ label, name }) => {
+        {tokenItems.map(({ label, name }) => {
           const isChecked = tokenTypes[name];
 
           return (
